[Fix] extra option validation
